HACK HACK HACK: Workarounds for binutils and fts
The DPACKAGE definition is needed to work around binutils wanting
to include config.h if the PACKAGE version isn't defined.

bpftool utilizes glibc fts functions.  Which doesn't exist in musl.
So we've made the bpftool package depend on musl-fts (which is a
shim fts implementation).  But there needs to be a better way of
providing the path to the fts archive than this.